Accept jam_date as optional argument in daily-jams patch
We want to create daily jams playlists in LB based on user timezones so
that those are ready by the time they wake up. This means that the patch
needs to take into account timezones when calculating the jam date to
display to users. One way to do that would be to accept a timezone
parameter and let the patch calculate the current day based on that. But
the other way to accept a jam date directly seems more flexible and
easier so doing that.