Add RecentListensTimestampLookup element (#72)
* Add RecentListensTimestampLookup element

We generate daily jams based on user's timezone in ListenBrainz. In many
cases, these jams will get generated before yesterday's listens get
imported in the Spark cluster. As a result the latest_listened_at timestamp,
of the recommendations can be out of date. To fix this, retrieve all
listens of the user from last 2 days and augument the latest_listened_at
value of the recordings with those.

* Add DAYS_OF_RECENT_LISTENS_TO_EXCLUDE constant