Use more defensive checks for ratelimits
Troi is used in LB radio in production where it can be invoked in an api request
context. Multiple api requests modifying the global domain_ratelimit_lookup
concurrently causes issues with the ratelimiting logic. For instance, one thread
can check that the key exists in the dict but another thread can delete the key
from the dict before the first has accessed the key leading to issues. Since, the
key is only deleted after the sleep it can be safely ignored by using more
defensive accesses.